+++ packages/mg-components/src/components/atoms/mg-character-left/test/mg-character-left.e2e.ts
@@ -1,17 +1,17 @@
-import { PageType, describe, expect, setPageContent, testEach } from '../../../../utils/playwright.e2e.test.utils';
+import { expect } from '@playwright/test';
+import { renderAttributes } from '@mgdis/playwright-helpers';
+import { test } from '../../../../utils/playwright.fixture';
 
-const TIMEOUT = 1000;
+test.describe('mg-character-left', () => {
+  [
+    { characters: '', maxlength: 100 },
+    { characters: 'blu', maxlength: 200 },
+    { characters: 'blu blu blu blu', maxlength: 1000 },
+  ].forEach((args, index) => {
+    test(`Should render ${index + 1}`, async ({ page }) => {
+      await page.setContent(`<mg-character-left ${renderAttributes(args)}></mg-character-left>`);
 
-describe('mg-character-left', () => {
-  testEach([
-    `<mg-character-left characters="" maxlength="100"></mg-character-left>`,
-    `<mg-character-left characters="blu" maxlength="200"></mg-character-left>`,
-    `<mg-character-left characters="blu blu blu blu" maxlength="1000"></mg-character-left>`,
-  ])('Should render %s', async (page: PageType, html: string) => {
-    await setPageContent(page, html);
-
-    await page.locator('mg-character-left.hydrated').waitFor({ timeout: TIMEOUT });
-
-    await expect(page.locator('.e2e-screenshot')).toHaveScreenshot();
+      await expect(page.locator('.e2e-screenshot')).toHaveScreenshot();
+    });
   });
 });