+++ packages/mg-components/src/components/atoms/mg-tooltip/test/mg-tooltip.spec.tsx
@@ -206,7 +206,7 @@ describe('mg-tooltip', () => {
       expect(tooltip).toHaveAttribute('data-show');
     });
 
-    test('should hide tooltip with keyboardEvent, case "Escape" code', async () => {
+    test('Should hide tooltip with keyboardEvent, case "Escape" code', async () => {
       const args = { identifier: 'identifier', message: 'batman' };
       const page = await getPage(args, <span id="batman">batman</span>);
       const mgTooltip = page.doc.querySelector('mg-tooltip');
@@ -338,7 +338,7 @@ describe('mg-tooltip', () => {
     expect(spy).toHaveBeenCalled();
   });
 
-  test('should update mg-popover-content id when "identifier" is updated', async () => {
+  test('Should update mg-popover-content id when "identifier" is updated', async () => {
     const page = await getPage(
       { identifier: 'identifier', message: 'My tooltip message' },
       <mg-button identifier="identifier" disabled>